修改netdata

+++ scripts/before_compile.sh
@@ -22,12 +22,13 @@ sed -i 's/KERNEL_PATCHVER:=5.15/KERNEL_PATCHVER:=6.1/g' target/linux/x86/Makefil
 sed -i '/CONFIG_NODEJS/d' .config
 
 # 替换netdata为中文版
-rm -rf ./feeds/luci/applications/luci-app-netdata
-git clone https://github.com/Jason6111/luci-app-netdata --delth=1 ./feeds/luci/applications/luci-app-netdata
+if grep -q "CONFIG_PACKAGE_luci-app-netdata=y" .config ; then
+    rm -rf ./feeds/luci/applications/luci-app-netdata
+    git clone https://github.com/Jason6111/luci-app-netdata --depth=1 ./feeds/luci/applications/luci-app-netdata
+fi
 
 # 如果用户配置了Nginx作为LUCI WEB服务器，则删除掉LUCI相关模块，以及调整NGINX的访问权限
 if grep -q "CONFIG_PACKAGE_luci-nginx=y" .config ; then
-    #sed -i 's/+uhttpd +uhttpd-mod-ubus //g' ./feeds/luci/collections/luci/Makefile
     sed -i '/uhttpd/d' .config
     sed -i '/deny/d' ./feeds/packages/net/nginx-util/files/restrict_locally
 fi
